add_two_numbers.py
given two linked list representing a number, where each node holds a digit ( in reverse order)

we add the values from each node, and denote them as sum and carry. the sum is inserted into the new node of the resultsmt lisst and the carry is added to the next node.